+++ packages/xprog-prensa/news/PageBlock/component.tsx
@@ -34,14 +34,31 @@ export const PageBlock: React.FC<PageBlockProps> = ({
   slotLayouts,
   slotTemplate,
   templates,
-  templateBgColor
+  templateBgColor,
+  wrappers
 }) => {
+  const getWrapperComponent = (wrapperFromProps, wrapperFromTemplate) => {
+    const wrapperTemplateEnabled = get(wrapperFromTemplate, 'enabled', false)
+    const wrapperPropsEnabled = get(wrapperFromProps, 'enabled', false)
+    if (wrapperTemplateEnabled) {
+      return wrapperFromTemplate
+    }
+    else if (wrapperPropsEnabled) {
+      return wrapperFromProps
+    }
+    else {
+      return undefined
+    }
+  }
   const getLayoutBySlotAndTemplate = (name, slot) => {
     const layoutSelected = get(slotLayouts, slot)
     const layoutSlotConfig = get(layoutSelected, `slotConfig[${slotTemplate}][slot${name}]`)
     return {
       css: layoutSelected?.css,
-      ...layoutSlotConfig
+      ...layoutSlotConfig,
+      wrapperTop: layoutSelected?.wrapperTop,
+      wrapperContent: layoutSelected?.wrapperContent,
+      wrapperBottom: layoutSelected?.wrapperBottom
     }
   }
   const leftColumnLayout = getLayoutBySlotAndTemplate('Left', slotLeftLayout)
@@ -109,6 +126,24 @@ export const PageBlock: React.FC<PageBlockProps> = ({
     itemComponent
   }
 
+  const columnExtraTopProps: ColumnProps = {
+    ...columnProps,
+    css: {
+      column: {
+        ...css?.column,
+        ...template?.css?.column
+      },
+      item: {
+        ...css?.item,
+        ...template?.css?.item
+      }
+    },
+    items: slotManualExtraTop,
+    name: 'ExtraTop',
+    wrapperTop: getWrapperComponent(wrappers?.columnExtraTop?.slotTop, leftColumnLayout.wrapperTop),
+    wrapperContent: getWrapperComponent(wrappers?.columnExtraTop?.slotMiddle ,leftColumnLayout.wrapperContent),
+    wrapperBottom: getWrapperComponent(wrappers?.columnExtraTop?.slotBottom ,leftColumnLayout.wrapperBottom)
+  }
   const columnLeftProps: ColumnProps = {
     ...columnProps,
     customProps: {
@@ -133,7 +168,10 @@ export const PageBlock: React.FC<PageBlockProps> = ({
     itemsBottom: slotManualBottomLeft,
     itemsTop: slotManualTopLeft,
     layout: leftColumnLayout,
-    name: 'Left'
+    name: 'Left',
+    wrapperTop: getWrapperComponent(wrappers?.columnLeft?.slotTop, leftColumnLayout.wrapperTop),
+    wrapperContent: getWrapperComponent(wrappers?.columnLeft?.slotMiddle ,leftColumnLayout.wrapperContent),
+    wrapperBottom: getWrapperComponent(wrappers?.columnLeft?.slotBottom ,leftColumnLayout.wrapperBottom)
   }
   const columnCenterProps: ColumnProps = {
     ...columnProps,
@@ -153,7 +191,10 @@ export const PageBlock: React.FC<PageBlockProps> = ({
     itemsBottom: slotManualBottomCenter,
     itemsTop: slotManualTopCenter,
     layout: centerColumnLayout,
-    name: 'Center'
+    name: 'Center',
+    wrapperTop: getWrapperComponent(wrappers?.columnCenter?.slotTop, centerColumnLayout.wrapperTop),
+    wrapperContent: getWrapperComponent(wrappers?.columnCenter?.slotMiddle, centerColumnLayout.wrapperContent),
+    wrapperBottom: getWrapperComponent(wrappers?.columnCenter?.slotBottom, centerColumnLayout.wrapperBottom)
   }
   const columnRightProps: ColumnProps = {
     ...columnProps,
@@ -173,7 +214,10 @@ export const PageBlock: React.FC<PageBlockProps> = ({
     itemsBottom: slotManualBottomRight,
     itemsTop: slotManualTopRight,
     layout: rightColumnLayout,
-    name: 'Right'
+    name: 'Right',
+    wrapperTop: getWrapperComponent(wrappers?.columnRight?.slotTop, rightColumnLayout.wrapperTop),
+    wrapperContent: getWrapperComponent(wrappers?.columnRight?.slotMiddle, rightColumnLayout.wrapperContent),
+    wrapperBottom: getWrapperComponent(wrappers?.columnRight?.slotBottom, rightColumnLayout.wrapperBottom)
   }
   const columnExtraProps: ColumnProps = {
     ...columnProps,
@@ -188,22 +232,10 @@ export const PageBlock: React.FC<PageBlockProps> = ({
       }
     },
     items: slotManualExtra,
-    name: 'Extra'
-  }
-  const columnExtraTopProps: ColumnProps = {
-    ...columnProps,
-    css: {
-      column: {
-        ...css?.column,
-        ...template?.css?.column
-      },
-      item: {
-        ...css?.item,
-        ...template?.css?.item
-      }
-    },
-    items: slotManualExtraTop,
-    name: 'ExtraTop'
+    name: 'Extra',
+    wrapperTop: getWrapperComponent(wrappers?.columnExtra?.slotTop, leftColumnLayout.wrapperTop),
+    wrapperContent: getWrapperComponent(wrappers?.columnExtra?.slotMiddle ,leftColumnLayout.wrapperContent),
+    wrapperBottom: getWrapperComponent(wrappers?.columnExtra?.slotBottom ,leftColumnLayout.wrapperBottom)
   }
 
   return (