+++ packages/xprog-prensa/primitives/ImageGallery/types.ts
@@ -1,3 +1,6 @@
+/* eslint-disable no-undef */
+import { CarouselProps } from '../Carousel/types'
+
 export type ItemProps = {
   caption: string;
   url: string;
@@ -6,11 +9,13 @@ export type ItemProps = {
 /**
  * Prensa | Image Gallery
  */
-export interface ImageGalleryProps {
+export interface ImageGalleryProps extends Omit<Omit<Omit<CarouselProps, 'labels'>, 'children'>, 'axis'> {
+  axis?: CarouselProps['axis'];
   captionProps?: any;
   componentCaption?: any;
   componentImage?: any;
   items: Array<ItemProps>;
+  labels?: CarouselProps['labels'];
   height?: string;
   width?: string;
 }